Refactor: Add comments coreml 타입에 따른 output 핸들링 구문 둘다 넣어놓기
