--- MilkGuide/MilkGuide/Camera/CameraViewController.swift
@@ -1,203 +1,0 @@
-//
-//  CameraViewController.swift
-//  MilkGuide
-//
-//  Created by 박희지 on 2021/06/26.
-//
-
-import UIKit
-import AVFoundation
-import Photos
-
-class CameraViewController: UIViewController {
-    let captureSession = AVCaptureSession()
-    var videoDeviceInput: AVCaptureDeviceInput! // 전∙후면 변경 가능
-    var photoOutput: AVCapturePhotoOutput = {
-        let output = AVCapturePhotoOutput()
-        // 고해상도 캡처 가능하게
-        output.isHighResolutionCaptureEnabled = true
-        return output
-    }()
-    
-    let sessionQueue = DispatchQueue(label: "session Queue")
-    let videoDeviceDiscoverySession = AVCaptureDevice.DiscoverySession(deviceTypes: [.builtInDualCamera, .builtInWideAngleCamera, .builtInTrueDepthCamera], mediaType: .video, position: .unspecified)
-    
-    @IBOutlet weak var previewView: previewView!
-    @IBOutlet weak var captureButton: UIButton!
-    @IBOutlet weak var blurBGView: UIVisualEffectView!
-    @IBOutlet weak var switchButton: UIButton!
-    
-    override var prefersStatusBarHidden: Bool {
-        return true
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // 초기 설정
-        previewView.session = captureSession
-        sessionQueue.async {
-            self.setupSession()
-            self.startSession()
-        }
-        setupUI()
-    }
-    
-    func setupUI() {
-        captureButton.layer.cornerRadius = captureButton.bounds.height/2
-        captureButton.layer.masksToBounds = true
-        
-        blurBGView.layer.cornerRadius = blurBGView.bounds.height/2
-        blurBGView.layer.masksToBounds = true
-    }
-    @IBAction func switchCamera(_ sender: Any) {
-        // 반대편 카메라 찾아서 재설정
-        // 1. 반대 카메라 찾기
-        // 2. 새로운 디바이스 가지고 세션 update
-        // 3. 카메라 토글 버튼 update
-        
-        sessionQueue.async {
-            let currentVideoDevice = self.videoDeviceInput.device
-            let currentPosition = currentVideoDevice.position
-            let isFront = currentPosition == .front
-            let preferredPosition: AVCaptureDevice.Position = isFront ? .back: .front
-            
-            let devices = self.videoDeviceDiscoverySession.devices
-            var newVideoDevice: AVCaptureDevice?
-            
-            newVideoDevice = devices.first(where: {device in
-                return preferredPosition == device.position
-            })
-            
-            // update capture session
-            if let newDevice = newVideoDevice {
-                do {
-                    let videoDeviceInput = try AVCaptureDeviceInput(device: newDevice)
-                    self.captureSession.beginConfiguration()
-                    self.captureSession.removeInput(self.videoDeviceInput)
-                    
-                    // add new device input
-                    if self.captureSession.canAddInput(videoDeviceInput) {
-                        self.captureSession.addInput(videoDeviceInput)
-                        self.videoDeviceInput = videoDeviceInput
-                    } else {
-                        self.captureSession.addInput(self.videoDeviceInput)
-                    }
-                    self.captureSession.commitConfiguration()
-                } catch let error {
-                    print("error occured while creating device input: \(error.localizedDescription)")
-                }
-            }
-        }
-    }
-    
-    @IBAction func captureCamera(_ sender: Any) {
-        // photoOutput의 capturePhoto 메소드
-        
-        let videoPreviewLayerOrientation =
-            self.previewView.videoPreviewLayer.connection?.videoOrientation
-        sessionQueue.async {
-            let connection = self.photoOutput.connection(with: .video)
-            connection?.videoOrientation = videoPreviewLayerOrientation!
-            let setting = AVCapturePhotoSettings()
-            self.photoOutput.capturePhoto(with: setting, delegate: self)    // photo output에게 지금 찍는다고 알려줌
-        }
-    }
-    
-    func savePhotoLibrary(image: UIImage) {
-        // TODO: capture한 이미지 포토라이브러리에 저장
-        // 라이브러리 권한 요청해야 함
-        PHPhotoLibrary.requestAuthorization { status in
-            if status == .authorized {
-                // 저장
-                PHPhotoLibrary.shared().performChanges({
-                    PHAssetChangeRequest.creationRequestForAsset(from: image)
-                }) { (success, error) in
-                    print("--> 이미지 저장 완료했나? \(success)")
-                }
-            } else {
-                // 다시 요청
-                print("--> 권한을 받지 못함")
-            }
-        }
-    }
-    
-}
-
-extension CameraViewController {
-    // MARK: - Setup session and preview
-    func setupSession() {
-        // 1. presentSetting
-        // 2. beginConfiguration
-        // 3. Add Video Input
-        // 4. Add Photo Output
-        // 5. commitConfiguration
-        
-        captureSession.sessionPreset = .high   // 해상도 설정
-        captureSession.beginConfiguration()
-        
-        // add video input
-        var defaultVideoDevice: AVCaptureDevice?
-        guard let camera = videoDeviceDiscoverySession.devices.first else {     // 디바이스 세션에서 기기찾아 가져오기
-            captureSession.commitConfiguration()
-            return
-        }
-        do {
-            let videoDeviceInput = try AVCaptureDeviceInput(device: camera)
-            if captureSession.canAddInput(videoDeviceInput) {
-                captureSession.addInput(videoDeviceInput)
-                self.videoDeviceInput = videoDeviceInput
-            } else {
-                captureSession.commitConfiguration()
-                return
-            }
-        } catch let error {
-            captureSession.commitConfiguration()
-            return
-        }
-        
-        // add photo output
-        photoOutput.setPreparedPhotoSettingsArray([AVCapturePhotoSettings(format: [AVVideoCodecKey: AVVideoCodecType.jpeg])], completionHandler: nil)   // 사진 저장 형식 설정
-        if captureSession.canAddOutput(photoOutput) {
-            captureSession.addOutput(photoOutput)
-        } else {
-            captureSession.commitConfiguration()
-            return
-        }
-        captureSession.commitConfiguration()
-    }
-    
-    func startSession() {
-        // session start
-        sessionQueue.async {
-            if !self.captureSession.isRunning {
-                self.captureSession.startRunning()
-            }
-        }
-    }
-    
-    func stopSession() {
-        // session stop
-        sessionQueue.async {
-            if self.captureSession.isRunning {
-                self.captureSession.stopRunning()
-            }
-        }
-    }
-}
-
-extension CameraViewController: AVCapturePhotoCaptureDelegate {
-    func photoOutput(_ output: AVCapturePhotoOutput, didFinishProcessingPhoto photo: AVCapturePhoto, error: Error?) {
-        guard error == nil else { return }
-        guard let imageData = photo.fileDataRepresentation() else { return }
-        // 이미지 라이브러리에 저장
-        guard let image = UIImage(data: imageData) else { return }
-        
-        let sb = UIStoryboard(name: "Main", bundle: nil)
-        let imageClassificationVC = sb.instantiateViewController(identifier: "ImageClassificationViewController") as! ImageClassificationViewController
-        imageClassificationVC.updateClassifications(for: image)
-        imageClassificationVC.modalPresentationStyle = .fullScreen
-        present(imageClassificationVC, animated: false, completion: nil)
-       
-        
-    }
-}
